Rename user defined subscriber behavior
