+++ dds/src/implementation/dds_impl/mod.rs
@@ -5,6 +5,7 @@ pub mod behavior_builtin_data_reader_stateful;
 pub mod behavior_builtin_data_reader_stateless;
 pub mod behavior_builtin_subscriber;
 pub mod behavior_user_defined_data_reader;
+pub mod behavior_user_defined_subscriber;
 pub mod behavior_user_defined_topic;
 pub mod dds_data_reader;
 pub mod dds_data_writer;
@@ -16,7 +17,6 @@ pub mod message_receiver;
 pub mod node_domain_participant;
 pub mod node_user_defined_data_writer;
 pub mod node_user_defined_publisher;
-pub mod node_user_defined_subscriber;
 pub mod nodes;
 pub mod participant_discovery;
 pub mod status_condition_impl;
